Use ERB instead of env for desktop
The form.yml specifies a value for the attribute "desktop"
this was being passed as an env var "DESKTOP" to the script.sh
to determine which script to run, gnome.sh or mate.sh

Instead, it is appropriate to use ERB here. This simplifies the code
and solves problems when it is difficult to pass environment variables
to the script, as in the case of LSF 8.3 when specifying -L "/bin/bash"
as arguments to bsub. In that case, the desired login shell is used, but
the environment used when submitting the job is wiped or ignored.